Simplify main run config to direct refs
