+++ pyproject.toml
@@ -14,9 +14,7 @@ dependencies = [
 
 [project.scripts]
 runmeta = "monocycle_nash.runmeta.cli:main"
-solve_payoff = "monocycle_nash.entrypoints.solve_payoff:main"
-graph_payoff = "monocycle_nash.entrypoints.graph_payoff:main"
-plot_characters = "monocycle_nash.entrypoints.plot_characters:main"
+main = "main:main"
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
