+++ src/monocycle_nash/equilibrium/__init__.py
@@ -1,3 +1,4 @@
+from .application import run_solve_payoff
 from .domain import MixedStrategy
 
-__all__ = ["MixedStrategy"]
+__all__ = ["MixedStrategy", "run_solve_payoff"]
