+++ src/monocycle_nash/loader/__init__.py
@@ -3,6 +3,14 @@
 """
 
 from .data_loader import ExperimentDataLoader, SettingDataLoader
+from .run_config import FeatureInputRefs, MAIN_RUN_CONFIG_PATH, RunConfigLoader
 from .toml_tree import TomlTreeLoader
 
-__all__ = ["TomlTreeLoader", "ExperimentDataLoader", "SettingDataLoader"]
+__all__ = [
+    "TomlTreeLoader",
+    "ExperimentDataLoader",
+    "SettingDataLoader",
+    "RunConfigLoader",
+    "FeatureInputRefs",
+    "MAIN_RUN_CONFIG_PATH",
+]
